backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Supozy Backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.12.3",
    "groq>=0.31.1",
    "ipykernel>=6.30.1",
    "mariadb>=1.1.13",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.1",
]
